Discussion: eBPF program & userspace program repo
This document is to define what a repository for eBPF programs and their
associated userspace programs should be and how it should operate.

A preview of this document was done in the Jan 5 2022 TSC meeting:
https://wiki.lfnetworking.org/display/L3AF/01-05-2022+TSC+Meeting+Minutes

We expect a lot of feedback and changes to this document; once it has
stabilized, it will move to a GitHub discussion as a proposal.

Signed-off-by: Santhosh Fernandes <santhosh.fernandes@gmail.com>